pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gallery-autosorter"
version = "0.1.0"
description = "Automatically organize photos into year/month folders based on EXIF metadata"
readme = "README.md"
requires-python = ">=3.7"
license = {text = "MIT"}
authors = [
    {name = "Guillermo", email = "your.email@example.com"}
]
keywords = ["photos", "organization", "exif", "metadata", "automation"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Graphics",
    "Topic :: Utilities",
]
dependencies = [
    "Pillow>=9.0.0",
    "piexif>=1.1.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "black>=22.0",
    "flake8>=4.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/gallery-autosorter"
Repository = "https://github.com/yourusername/gallery-autosorter.git"
Issues = "https://github.com/yourusername/gallery-autosorter/issues"

[tool.setuptools]
packages = ["src"]

[tool.black]
line-length = 100
target-version = ["py37"]

[tool.isort]
profile = "black"
line_length = 100
